+++ tests/TestCase.php
@@ -27,6 +27,7 @@ use Google\Cloud\Spanner\Date;
 use Google\Cloud\Spanner\Numeric;
 use Google\Cloud\Spanner\SpannerClient;
 use Illuminate\Foundation\Application;
+use Illuminate\Support\Carbon;
 use Ramsey\Uuid\Uuid;
 
 /**
@@ -59,7 +60,7 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
 
     protected function generateTableName(string $prefix = 'Temp'): string
     {
-        return $prefix . '_' . date('Ymd_His_v');
+        return $prefix . '_' . Carbon::now()->format('Ymd_His_v');
     }
 
     /**