+++ CHANGELOG.md
@@ -9,8 +9,9 @@ Added
 - Allow `spanner_emulator.disable_query_null_filtered_index_check` to be set (#180)
 - Allow default max transaction attempts to be changed (#179)
 - Table prefixing support (#172)
-- Support for GENERATE_UUID() in migrations (#174)
+- Support for `GENERATE_UUID()` in migrations (#174)
 - Support `Schema\Blueprint::increments()` by using a column of type `STRING(36)` with `DEFAULT (GENERATE_UUID())` (#175)
+- Support for `CREATE SEQUENCE` in migrations (#181)
 
 Changed
 - `Query\Builder::lock()` no longer throw an error and will be ignored instead (#156)